refactor: generalize FilterBottomSheetHeader to BottomSheetHeader
